backend/requirements-dev.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-decouple==3.8
alembic==1.12.1
pydantic[email]==2.6.0
pydantic-settings==2.2.1
pytest==7.4.3
httpx==0.25.2

# Cache & Performance
redis==5.0.1
aioredis==2.0.1

# Monitoring & Analytics
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

# Sécurité avancée
bcrypt==4.0.1
cryptography==41.0.7
pyjwt[crypto]==2.8.0

# Validation avancée
email-validator==2.1.0
pydantic-extra-types==2.3.0

# Tâches asynchrones
celery==5.3.4
flower==2.0.1

# Tests avancés
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Documentation API
mkdocs==1.5.3
mkdocs-material==9.4.8
